ActionGraphParser moved to resolvers
